mysite/requirements.txt
backports.ssl-match-hostname==3.5.0.1
cffi==1.6.0
cryptography==1.3.1
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
docker-py==1.8.0
enum34==1.1.4
idna==2.1
ipaddress==1.0.16
Markdown==2.6.6
ndg-httpsclient==0.4.0
py2-ipaddress==3.4.1
pyasn1==0.1.9
pycparser==2.14
PyMySQL==0.7.2
pyOpenSSL==16.0.0
requests==2.9.1
six==1.10.0
websocket-client==0.37.0
